Added gamification lesson for day 2
Basically talks about how to design with humans in mind. For this I feel I realized that making something entertaining means looking at what the consumers desire most and then giving them that, which means giving them a way of making the meaning they want more of.
Look around you. What do you see that is function-focused?
How can it be improved by focusing on solving the need of the human as opposed to the need
to complete a task?